Get rid of hidpi option, add credits
The way hidpi was done was a bit hacky since wayland's fractional
scaling used to be quite a mess and what you would do is set the
variables controlling the individual toolkits's scaling size to
fit the highest resolution monitor you had and then decrease
that monitor's scale factor to have it of appropriate size.
The downside of this approach was that for each app you would
need to find out and adjust the scaling settings (if those were
even present). It used to be a mess and after more than 2 months
of driving the proper wayland scaling approach without any hicckups
I've decided to delete this functionality. It also had a fixed scaling
size, mostly fit for 2k monitors, now the users can more flexibly
select the size just be modifying their kanshi/dwl configs. The
only downside of this approach is that it makes xwayland apps
look blurry and xwayland games (using wine) problematic but
since marbs is not inteded for gamers, this should not be an issue